CommandFiles agora está dentro de routes.js
