+++ config.json
@@ -1,3 +1,3 @@
 {
-  "BOT_TOKEN": "MTA1NzMxOTc4NTk1Nzk3MDA4MA.G4qTRP.MUXfixpugL1VOBtbwjsvTA3_7tbzUQcNZZ_IC8"
+  "BOT_TOKEN": "MTA1NzMxOTc4NTk1Nzk3MDA4MA.GnTgdh.hUD1uNA1BtNijMMEfLsO4f1IMfqsySaXzcxEKE"
 }
